refactor(policy-graph): split evaluator condition helpers
Move the condition, evidence, predicate, and edge-matching helpers out of evaluation.rs into evaluation_conditions.rs to keep the evaluator under the source-length cap after the approval-gate additions. No behavior change.
